+++ CHANGELOG.md
@@ -17,6 +17,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 - More accurate APU frame interrupt flag
 - Fixed an inaccuracy in the controller input implementation
 - More accurate PPU implementation
+- More accurate sprite evaluation
 
 ## [0.20.0] - 2026-09-16